+++ yandex/data_source_yandex_mdb_redis_cluster.go
@@ -62,6 +62,14 @@ func dataSourceYandexMDBRedisCluster() *schema.Resource {
 							Type:     schema.TypeInt,
 							Computed: true,
 						},
+						"client_output_buffer_limit_normal": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"client_output_buffer_limit_pubsub": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
 						"databases": {
 							Type:     schema.TypeInt,
 							Computed: true,
@@ -250,13 +258,15 @@ func dataSourceYandexMDBRedisClusterRead(d *schema.ResourceData, meta interface{
 	conf := extractRedisConfig(cluster.Config)
 	err = d.Set("config", []map[string]interface{}{
 		{
-			"timeout":                 conf.timeout,
-			"maxmemory_policy":        conf.maxmemoryPolicy,
-			"version":                 conf.version,
-			"notify_keyspace_events":  conf.notifyKeyspaceEvents,
-			"slowlog_log_slower_than": conf.slowlogLogSlowerThan,
-			"slowlog_max_len":         conf.slowlogMaxLen,
-			"databases":               conf.databases,
+			"timeout":                           conf.timeout,
+			"maxmemory_policy":                  conf.maxmemoryPolicy,
+			"version":                           conf.version,
+			"notify_keyspace_events":            conf.notifyKeyspaceEvents,
+			"slowlog_log_slower_than":           conf.slowlogLogSlowerThan,
+			"slowlog_max_len":                   conf.slowlogMaxLen,
+			"databases":                         conf.databases,
+			"client_output_buffer_limit_normal": conf.clientOutputBufferLimitNormal,
+			"client_output_buffer_limit_pubsub": conf.clientOutputBufferLimitPubsub,
 		},
 	})
 	if err != nil {