+++ CHANGELOG.md
@@ -7,6 +7,7 @@ BUG FIXES:
   
 ENHANCEMENTS:
 * alb: refactoring `yandex_alb_backend_group` resource and data source
+* mdb: add `client_output_buffer_limit_normal` and `client_output_buffer_limit_pubsub` attributes to `yandex_mdb_redis_cluster` resource and data source
 
 FEATURES:
 * compute: support for creating `yandex_compute_instance` with local disks